Split wf module
